+++ R/assertthat.R
@@ -79,8 +79,8 @@ on_failure(is_counts_vector) <- function(call, env) {
 #' is_list_with(b, c("a", "c"))
 #' is_list_with(b, c("a", "b"))
 is_list_with <- function(x, elements) {
-  assert_that(utils.nest::is_character_vector(elements))
-  utils.nest::is_fully_named_list(x) &&
+  assert_character(elements, any.missing = FALSE, min.len = 1L)
+  test_list(x, names = "unique") &&
     all(elements %in% names(x))
 }