+++ R/draw_barplot.R
@@ -75,10 +75,7 @@ draw_barplot <- function(object,
 
   col_data <- colData(object)
 
-  percentiles_without_borders <- setdiff(percentiles, c(0, 1))
-  df <- data.frame(
-    x = tern::cut_quantile_bins(x_cont, probs = percentiles_without_borders)
-  )
+  df <- data.frame(x = cut_quantile(x_cont, percentiles))
 
   if (!is.null(facet_var)) {
     assert_names(names(col_data), must.include = facet_var)