+++ R/normalization.R
@@ -104,16 +104,24 @@ setMethod(
                         methods = c("cpm", "rpkm", "tpm", "voom", "vst"),
                         control = control_normalize(),
                         ...) {
-    method_choices <- c("cpm", "rpkm", "tpm", "voom", "vst", "rlog")
-    assert_that(all(methods %in% method_choices))
+    norm_funs <- list(
+      cpm = h_cpm,
+      rpkm = h_rpkm,
+      tpm = h_tpm,
+      voom = h_voom,
+      vst = h_vst,
+      rlog = h_rlog
+    )
+    method_choices <- names(norm_funs)
+    assert_subset(x = methods, choices = method_choices)
     methods <- match.arg(methods, choices = method_choices, several.ok = TRUE)
+
     for (method in methods) {
-      fun_name <- paste0("hermes::h_", method)
-      method_result <- eval(utils.nest::call_with_colon(
-        fun_name,
+      fun <- norm_funs[[method]]
+      method_result <- fun(
         object = object,
         control = control
-      ))
+      )
       assay(object, method) <- method_result
     }
     metadata(object) <- c(metadata(object), list(control_normalize = control))